update ipex_accelerator and ipex_strategy
