+++ python/nano/setup.py
@@ -72,7 +72,7 @@ def setup_package():
 
     pytorch_requires = ["torch==1.11.0",
                         "torchvision==0.12.0",
-                        "pytorch_lightning==1.4.2",
+                        "pytorch_lightning==1.6.4",
                         "torchmetrics==0.7.2",
                         "opencv-python-headless",
                         "PyTurboJPEG",
