+++ python/nano/src/bigdl/nano/deps/ipex/ipex_strategy.py
@@ -1,0 +1,66 @@
+#
+# Copyright 2016 The BigDL Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from typing import Optional
+
+import torch
+import pytorch_lightning as pl
+from pytorch_lightning.strategies import SingleDeviceStrategy
+from pytorch_lightning.accelerators.accelerator import Accelerator
+from pytorch_lightning.plugins.precision import PrecisionPlugin
+from bigdl.nano.utils.log4Error import invalidInputError
+import intel_extension_for_pytorch as ipex
+
+from .ipex_accelerator import IPEXAccelerator
+
+class IPEXStrategy(SingleDeviceStrategy):
+
+    def __init__(
+        self,
+        accelerator: Accelerator = IPEXAccelerator(),
+        precision_plugin: PrecisionPlugin = PrecisionPlugin(),
+        enable_bf16=False,
+    ) -> None:
+        """
+        Args:
+            accelerator: the accelerator to handle hardware
+            precision_plugin: the plugin to handle precision-specific parts
+        """
+        self.enable_bf16 = enable_bf16
+        if enable_bf16:
+            # Automatically mix precision
+            ipex.enable_auto_mixed_precision(mixed_dtype=torch.bfloat16)
+
+        super().__init__(accelerator=accelerator, precision_plugin=precision_plugin)
+
+    def setup(self, trainer: pl.Trainer, *args, **kwargs) -> None:
+        """
+        Setup plugins for the trainer fit and creates optimizers.
+        Args:
+            trainer: the trainer instance
+            kwargs['model']: pl.LightningModule
+        """
+
+        self.setup_precision_plugin()
+        self.setup_optimizers(trainer)
+        model = kwargs['model']
+
+        if len(self.optimizers) > 1:
+            invalidInputError(False, "Ipex does not support more than one optimizers.")
+        dtype = torch.bfloat16 if self.enable_bf16 else None
+        model, optimizer = ipex.optimize(model, optimizer=self.optimizers[0],
+                                         inplace=True, dtype=dtype)
+        self.optimizers = [optimizer]
+
