+++ python/nano/src/bigdl/nano/deps/ipex/version_1_9/ipex_accelerator_1_9.py
@@ -14,102 +14,33 @@
 # limitations under the License.
 #
 
-from typing import Union, Dict, Any, Optional
-import logging
-import intel_pytorch_extension as ipex
+from typing import Any
 
-import torch
-from torch.optim import Optimizer
-
-import pytorch_lightning as pl
-from pytorch_lightning.plugins import SingleDevicePlugin
 from pytorch_lightning.accelerators.accelerator import Accelerator
-from pytorch_lightning.plugins.training_type import TrainingTypePlugin
-from pytorch_lightning.plugins.precision import PrecisionPlugin, MixedPrecisionPlugin
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from bigdl.nano.utils.log4Error import invalidInputError
-
-from .ipex_torchfunctional import to_cpu
-
-_STEP_OUTPUT_TYPE = Union[torch.Tensor, Dict[str, Any]]
 
 
 class IPEXAccelerator(Accelerator):
     """ Accelerator for XPU devices. """
 
-    def __init__(
-        self,
-        precision_plugin: PrecisionPlugin = PrecisionPlugin(),
-        training_type_plugin: TrainingTypePlugin = SingleDevicePlugin(
-            torch.device(ipex.DEVICE)),
-        enable_bf16=False,
-    ) -> None:
-        """
-        Args:
-            precision_plugin: the plugin to handle precision-specific parts
-            training_type_plugin: the plugin to handle different training routines
-        """
-        if enable_bf16:
-            # Automatically mix precision
-            ipex.enable_auto_mixed_precision(mixed_dtype=torch.bfloat16)
+    def __init__(self) -> None:
+        super().__init__()
+    
+    @staticmethod
+    def is_available() -> bool:
+        """Detect if IPEX accelerator is available"""
+        return True
 
-        self.device = ipex.DEVICE
+    @staticmethod
+    def auto_device_count() -> int:
+        """Get the devices when set to auto."""
+        return 1
 
-        super().__init__(precision_plugin=precision_plugin,
-                         training_type_plugin=training_type_plugin)
+    @staticmethod
+    def parse_devices(devices: Any) -> Any:
+        """Accelerator device parsing logic."""
+        pass
 
-    def setup(self, trainer: 'pl.Trainer', model: 'pl.LightningModule') -> None:
-        """
-        throw:
-            MisconfigurationException:
-                If AMP is used with XPU.
-        """
-        if isinstance(self.precision_plugin, MixedPrecisionPlugin):
-            invalidInputError(False,
-                              "amp is not supported in bigdl-nano.")
-
-        return super().setup(trainer, model)
-
-    def training_step_end(self, output: _STEP_OUTPUT_TYPE) -> _STEP_OUTPUT_TYPE:
-        """
-        For ipex xpu tensor do not support `tensor.storage()` right now,
-        which is a required operation by pytorch_lightning,
-        so just move output to cpu to store it, and move it back when doing backward.
-        """
-        output = to_cpu(output)
-
-        return super().training_step_end(output)
-
-    def test_step_end(self, output: Optional[_STEP_OUTPUT_TYPE]) -> \
-            Optional[_STEP_OUTPUT_TYPE]:
-        """A hook to do something at the end of the test step
-        Args:
-            output: the output of the test step
-        """
-        output = to_cpu(output)
-
-        return super().test_step_end(output)
-
-    def validation_step_end(self, output: Optional[_STEP_OUTPUT_TYPE]) -> \
-            Optional[_STEP_OUTPUT_TYPE]:
-        """A hook to do something at the end of the validation step
-        Args:
-            output: the output of the validation step
-        """
-        output = to_cpu(output)
-
-        return super().validation_step_end(output)
-
-    def backward(self,  # type: ignore
-                 closure_loss: torch.Tensor,
-                 *args,
-                 **kwargs) -> torch.Tensor:
-        """
-        Moving back loss to xpu device
-        """
-        closure_loss = closure_loss.to(self.device)
-        return super().backward(
-            closure_loss,
-            *args,
-            **kwargs,
-        )
+    @staticmethod
+    def get_parallel_devices(devices: Any) -> Any:
+        """Gets parallel devices for the Accelerator."""
+        pass
